Add files via upload
Uses built in biome defs against lib_materials nodes.  Nodes can be pointed to default or other.
Removes commented code sections and refactors the on_generated function to be more efficient.
Biomes are not actually defined, but heat and humidity are used to determine the top node, filler and stone.

TODO: edit this dev code to use default mod, instead of lib_materials / lib_ecology.  This should ensure that it can run for most users.